Add files via upload
Using advanced CSS techniques, create a microinteraction and tutorial then demonstrate it to the class.

Microinteraction
Focus on the animation and animation techniques, especially multiple keyframes. Also include at least one of the following: transforms, gradients, filters. Be sure to include the "rule" and  "feedback" parts of the microinteraction.

The "trigger" and "loop/mode" parts can be described in the tutorial and demonstration but do not have to be implemented.

Tutorial
A write up with an introduction, step-by-step guide on your practical implementation, and a summary. Describe the "trigger" and "loop/mode" parts of the microinteraction.

Demonstration
Slides are not required. Your presentation to the class should be 5 minutes long and should consist of your functioning microinteraction and an explanation of it. The explanation should include the the "trigger" and "loop/mode" parts of the microinteraction.